댓글삭제기능,좋아요기능

+++ articles/views.py
@@ -84,8 +84,19 @@ def comment_create(request, pk):
     return redirect("articles:detail", article.pk)
 
 
-@login_required
-def comments_delete(request, article_pk, comment_pk):
+def comment_delete(request, pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
-    return redirect("articles:detail", article_pk, comment_pk)
+    if request.user == comment.user:
+        comment.delete()
+        messages.warning(request, "댓글이 삭제되었습니다.")
+    return redirect("articles:detail", pk)
+
+
+@login_required
+def like(request, pk):
+    article = Article.objects.get(pk=pk)
+    if request.user in article.like_users.all():
+        article.like_users.remove(request.user)
+    else:
+        article.like_users.add(request.user)
+    return redirect("articles:detail", pk)
